Removal of dangerous and deprecated API in directories.c
MSDN documentation says:
"SHGetMalloc was introduced in Windows 95 and Microsoft Windows NT 4.0, but as of Windows 2000 it is no longer necessary. "
Removed for the sake of my head, a WIN95 API still here....
IMalloc is also dangerous and luckily it was related to that API, so can go away too.
Let me know if it works ok. Tested on my build all is fine.